+++ src/main/java/in/koreatech/koin/domain/owner/controller/SmsBasedOwnerController.java
@@ -1,0 +1,96 @@
+package in.koreatech.koin.domain.owner.controller;
+
+import in.koreatech.koin.domain.owner.dto.OwnerAccountCheckExistsRequest;
+import in.koreatech.koin.domain.owner.dto.OwnerLoginRequest;
+import in.koreatech.koin.domain.owner.dto.OwnerLoginResponse;
+import in.koreatech.koin.domain.owner.dto.OwnerPasswordResetVerifySmsRequest;
+import in.koreatech.koin.domain.owner.dto.OwnerPasswordUpdateSmsRequest;
+import in.koreatech.koin.domain.owner.dto.OwnerRegisterByPhoneRequest;
+import in.koreatech.koin.domain.owner.dto.OwnerSendSmsRequest;
+import in.koreatech.koin.domain.owner.dto.OwnerSmsVerifyRequest;
+import in.koreatech.koin.domain.owner.dto.OwnerVerifyResponse;
+import in.koreatech.koin.domain.owner.dto.VerifySmsRequest;
+import in.koreatech.koin.domain.owner.service.OwnerSmsVerificationService;
+import jakarta.validation.Valid;
+import java.net.URI;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class SmsBasedOwnerController implements SmsBasedOwnerApi {
+
+    private final OwnerSmsVerificationService ownerSmsService;
+
+    @PostMapping("/owner/login")
+    public ResponseEntity<OwnerLoginResponse> ownerLogin(
+            @RequestBody @Valid OwnerLoginRequest request
+    ) {
+        OwnerLoginResponse response = ownerSmsService.ownerLogin(request);
+        return ResponseEntity.created(URI.create("/"))
+                .body(response);
+    }
+
+    @PostMapping("/owners/verification/sms")
+    public ResponseEntity<Void> requestVerificationToRegisterBySms(
+        @RequestBody @Valid VerifySmsRequest verifySmsRequest
+    ) {
+        ownerSmsService.requestSignUpSmsVerification(verifySmsRequest);
+        return ResponseEntity.ok().build();
+    }
+
+    @PostMapping("/owners/register/phone")
+    public ResponseEntity<Void> registerByPhone(
+        @Valid @RequestBody OwnerRegisterByPhoneRequest request
+    ) {
+        ownerSmsService.registerByPhone(request);
+        return ResponseEntity.ok().build();
+    }
+
+    @PostMapping("/owners/verification/code/sms")
+    public ResponseEntity<OwnerVerifyResponse> codeVerification(
+        @Valid @RequestBody OwnerSmsVerifyRequest request
+    ) {
+        OwnerVerifyResponse response = ownerSmsService.verifyCodeBySms(request);
+        return ResponseEntity.ok().body(response);
+    }
+
+    @PostMapping("/owners/password/reset/verification/sms")
+    public ResponseEntity<Void> sendResetPasswordBySms(
+        @Valid @RequestBody OwnerSendSmsRequest request
+    ) {
+        ownerSmsService.sendResetPasswordBySms(request);
+        return ResponseEntity.ok().build();
+    }
+
+    @PostMapping("/owners/password/reset/send/sms")
+    public ResponseEntity<Void> sendVerifyCodeBySms(
+        @Valid @RequestBody OwnerPasswordResetVerifySmsRequest request
+    ) {
+        ownerSmsService.verifyResetPasswordCodeBySms(request);
+        return ResponseEntity.ok().build();
+    }
+
+    @PutMapping("/owners/password/reset/sms")
+    public ResponseEntity<Void> updatePasswordBySms(
+        @Valid @RequestBody OwnerPasswordUpdateSmsRequest request
+    ) {
+        ownerSmsService.updatePasswordBySms(request);
+        return ResponseEntity.ok().build();
+    }
+
+    @GetMapping("/owners/exists/account")
+    public ResponseEntity<Void> checkDuplicationOfPhoneNumber(
+        @ModelAttribute("account")
+        @Valid OwnerAccountCheckExistsRequest request
+    ) {
+        ownerSmsService.checkExistsAccount(request);
+        return ResponseEntity.ok().build();
+    }
+}
