+++ src/main/java/in/koreatech/koin/domain/owner/service/OwnerSmsVerificationService.java
@@ -1,0 +1,185 @@
+package in.koreatech.koin.domain.owner.service;
+
+import static in.koreatech.koin.domain.user.model.UserType.OWNER;
+
+import in.koreatech.koin.domain.owner.dto.CompanyNumberCheckRequest;
+import in.koreatech.koin.domain.owner.dto.OwnerAccountCheckExistsRequest;
+import in.koreatech.koin.domain.owner.dto.OwnerEmailVerifyRequest;
+import in.koreatech.koin.domain.owner.dto.OwnerLoginRequest;
+import in.koreatech.koin.domain.owner.dto.OwnerLoginResponse;
+import in.koreatech.koin.domain.owner.dto.OwnerPasswordResetVerifyEmailRequest;
+import in.koreatech.koin.domain.owner.dto.OwnerPasswordResetVerifySmsRequest;
+import in.koreatech.koin.domain.owner.dto.OwnerPasswordUpdateEmailRequest;
+import in.koreatech.koin.domain.owner.dto.OwnerPasswordUpdateSmsRequest;
+import in.koreatech.koin.domain.owner.dto.OwnerRegisterByPhoneRequest;
+import in.koreatech.koin.domain.owner.dto.OwnerRegisterRequest;
+import in.koreatech.koin.domain.owner.dto.OwnerSendEmailRequest;
+import in.koreatech.koin.domain.owner.dto.OwnerSendSmsRequest;
+import in.koreatech.koin.domain.owner.dto.OwnerSmsVerifyRequest;
+import in.koreatech.koin.domain.owner.dto.OwnerVerifyResponse;
+import in.koreatech.koin.domain.owner.dto.VerifyEmailRequest;
+import in.koreatech.koin.domain.owner.dto.VerifySmsRequest;
+import in.koreatech.koin.domain.owner.exception.DuplicationCompanyNumberException;
+import in.koreatech.koin.domain.owner.exception.DuplicationPhoneNumberException;
+import in.koreatech.koin.domain.owner.model.Owner;
+import in.koreatech.koin.domain.owner.model.OwnerEmailRequestEvent;
+import in.koreatech.koin.domain.owner.model.OwnerRegisterBySmsEvent;
+import in.koreatech.koin.domain.owner.model.OwnerRegisterEvent;
+import in.koreatech.koin.domain.owner.model.OwnerShop;
+import in.koreatech.koin.domain.owner.model.OwnerSmsRequestEvent;
+import in.koreatech.koin.domain.owner.model.redis.DailyVerificationLimit;
+import in.koreatech.koin.domain.owner.model.redis.OwnerVerificationStatus;
+import in.koreatech.koin.domain.owner.repository.OwnerRepository;
+import in.koreatech.koin.domain.owner.repository.OwnerShopRedisRepository;
+import in.koreatech.koin.domain.owner.repository.redis.DailyVerificationLimitRepository;
+import in.koreatech.koin.domain.owner.repository.redis.OwnerVerificationStatusRepository;
+import in.koreatech.koin.domain.shop.model.shop.Shop;
+import in.koreatech.koin.domain.shop.repository.shop.ShopRepository;
+import in.koreatech.koin.domain.user.model.User;
+import in.koreatech.koin.domain.user.model.UserToken;
+import in.koreatech.koin.domain.user.repository.UserRepository;
+import in.koreatech.koin.domain.user.repository.UserTokenRepository;
+import in.koreatech.koin.global.auth.JwtProvider;
+import in.koreatech.koin.global.auth.exception.AuthorizationException;
+import in.koreatech.koin.global.domain.email.exception.DuplicationEmailException;
+import in.koreatech.koin.global.domain.email.form.OwnerRegistrationData;
+import in.koreatech.koin.global.domain.email.service.MailService;
+import in.koreatech.koin.global.domain.random.model.CertificateNumberGenerator;
+import in.koreatech.koin.global.exception.KoinIllegalArgumentException;
+import in.koreatech.koin.global.naver.service.NaverSmsService;
+import java.time.LocalDateTime;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.UUID;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class OwnerSmsVerificationService {
+
+    private final JwtProvider jwtProvider;
+    private final UserTokenRepository userTokenRepository;
+    private final UserRepository userRepository;
+    private final PasswordEncoder passwordEncoder;
+    private final ApplicationEventPublisher eventPublisher;
+    private final OwnerVerificationStatusRepository ownerVerificationStatusRepository;
+    private final DailyVerificationLimitRepository dailyVerificationLimitRedisRepository;
+    private final OwnerRepository ownerRepository;
+    private final ShopRepository shopRepository;
+    private final OwnerShopRedisRepository ownerShopRedisRepository;
+    private final NaverSmsService naverSmsService;
+
+    @Transactional
+    public OwnerLoginResponse ownerLogin(OwnerLoginRequest request) {
+        Owner owner = ownerRepository.getByAccount(request.account());
+        User user = owner.getUser();
+
+        if (!user.isSamePassword(passwordEncoder, request.password())) {
+            throw new KoinIllegalArgumentException("비밀번호가 틀렸습니다.");
+        }
+
+        if (!user.isAuthed()) {
+            throw new AuthorizationException("미인증 상태입니다. 인증을 진행해주세요.");
+        }
+
+        String accessToken = jwtProvider.createToken(user);
+        String refreshToken = String.format("%s-%d", UUID.randomUUID(), user.getId());
+        UserToken savedToken = userTokenRepository.save(UserToken.create(user.getId(), refreshToken));
+        user.updateLastLoggedTime(LocalDateTime.now());
+
+        return OwnerLoginResponse.of(accessToken, savedToken.getRefreshToken());
+    }
+
+    @Transactional
+    public void registerByPhone(OwnerRegisterByPhoneRequest request) {
+        if (userRepository.findByPhoneNumberAndUserType(request.phoneNumber(), OWNER).isPresent()) {
+            throw DuplicationPhoneNumberException.withDetail("account: " + request.phoneNumber());
+        }
+        if (ownerRepository.findByCompanyRegistrationNumber(request.companyNumber()).isPresent()) {
+            throw DuplicationCompanyNumberException.withDetail("companyNumber: " + request.companyNumber());
+        }
+        Owner owner = request.toOwner(passwordEncoder);
+        Owner saved = ownerRepository.save(owner);
+        OwnerShop.OwnerShopBuilder ownerShopBuilder = OwnerShop.builder().ownerId(owner.getId());
+        if (request.shopId() != null) {
+            Shop shop = shopRepository.getById(request.shopId());
+            ownerShopBuilder.shopId(shop.getId());
+        }
+        ownerShopRedisRepository.save(ownerShopBuilder.build());
+        eventPublisher.publishEvent(new OwnerRegisterBySmsEvent(saved));
+    }
+
+    @Transactional
+    public void requestSignUpSmsVerification(VerifySmsRequest request) {
+        userRepository.findByPhoneNumberAndUserType(request.phoneNumber(), OWNER).ifPresent(user -> {
+            throw DuplicationPhoneNumberException.withDetail("account: " + request.phoneNumber());
+        });
+        sendCertificationSms(request.phoneNumber());
+    }
+
+    @Transactional
+    public OwnerVerifyResponse verifyCodeBySms(OwnerSmsVerifyRequest request) {
+        verifyCode(request.phoneNumber(), request.certificationCode());
+        String token = jwtProvider.createTemporaryToken();
+        return new OwnerVerifyResponse(token);
+    }
+
+    @Transactional
+    public void sendResetPasswordBySms(OwnerSendSmsRequest request) {
+        User user = userRepository.getByPhoneNumber(request.phoneNumber(), OWNER);
+        sendCertificationSms(user.getPhoneNumber());
+    }
+
+    @Transactional
+    public void verifyResetPasswordCodeBySms(OwnerPasswordResetVerifySmsRequest request) {
+        verifyCode(request.phoneNumber(), request.certificationCode());
+    }
+
+    @Transactional
+    public void updatePasswordBySms(OwnerPasswordUpdateSmsRequest request) {
+        User user = userRepository.getByPhoneNumber(request.phoneNumber(), OWNER);
+        user.updatePassword(passwordEncoder, request.password());
+    }
+
+    private void setVerificationCount(String key) {
+        Optional<DailyVerificationLimit> dailyVerificationLimit = dailyVerificationLimitRedisRepository.findById(key);
+        if (dailyVerificationLimit.isEmpty()) {
+            dailyVerificationLimitRedisRepository.save(new DailyVerificationLimit(key));
+        } else {
+            DailyVerificationLimit dailyVerification = dailyVerificationLimit.get();
+            dailyVerification.requestVerification();
+            dailyVerificationLimitRedisRepository.save(dailyVerification);
+        }
+    }
+
+    private void sendCertificationSms(String phoneNumber) {
+        setVerificationCount(phoneNumber);
+        String certificationCode = CertificateNumberGenerator.generate();
+        naverSmsService.sendVerificationCode(certificationCode, phoneNumber);
+        OwnerVerificationStatus ownerVerificationStatus = new OwnerVerificationStatus(
+            phoneNumber,
+            certificationCode
+        );
+        ownerVerificationStatusRepository.save(ownerVerificationStatus);
+        eventPublisher.publishEvent(new OwnerSmsRequestEvent(phoneNumber));
+    }
+
+    private void verifyCode(String key, String code) {
+        OwnerVerificationStatus verify = ownerVerificationStatusRepository.getByVerify(key);
+        if (!Objects.equals(verify.getCertificationCode(), code)) {
+            throw new KoinIllegalArgumentException("인증번호가 일치하지 않습니다.");
+        }
+        ownerVerificationStatusRepository.deleteById(key);
+    }
+
+    public void checkExistsAccount(OwnerAccountCheckExistsRequest request) {
+        ownerRepository.findByAccount(request.account()).ifPresent(user -> {
+            throw DuplicationPhoneNumberException.withDetail("account: " + request.account());
+        });
+    }
+}
